+++ src/Pages/EditEvent.css
@@ -1,21 +1,3 @@
-.editEvent__box {
-  align-items: center;
-  margin-top: 3rem;
-  margin-left: 4rem;
-  margin-right: 4rem;
-}
 
-.editEvent__save__button {
-  background-color: #ff355e !important;
-  border: #ff355e !important;
-  border-radius: 50px !important;
-  text-align: center !important;
-  font-weight: bold !important;
-  font-size: 18px;
-  height: 3rem;
 
-  text-align: center;
-}
 
-.form-label__editEvent {
-}
