+++ src/Pages/EditEvent.js
@@ -1,95 +1,13 @@
-import React, { useState } from 'react'
+import React from 'react'
 import './EditEvent.css'
 import CalendarHeader from '../Components/CalendarHeader.js'
-import 'react-datepicker/dist/react-datepicker.css'
-import DatePicker from 'react-datepicker'
-import { useLocation } from 'react-router-dom'
-import { useNavigate } from 'react-router-dom'
-import Button from 'react-bootstrap/Button'
+import EditOptions from '../Components/EventOptions.js'
 
-const EditEvent = (props) => {
-  const navigate = useNavigate()
-  const redirection = () => {
-    navigate('/', {
-      state: {
-        editEvents: props.selectedEvent,
-      },
-    })
-  }
-  const location = useLocation()
-
+const EditEvent = () => {
   return (
     <div>
       <CalendarHeader />
-      <div className="editEvent__box">
-        <form className="row g-3">
-          <div className="col-md-6">
-            <label for="inputEmail4" className="form-label">
-              Title
-            </label>
-            <input
-              type="text"
-              className="form-control"
-              id="inputEmail4"
-              defaultValue={location.state.editEvents.title}
-            />
-          </div>
-          <div className="col-md-6">
-            <label for="inputPassword4" className="form-label">
-              Description
-            </label>
-            <input
-              type="text"
-              className="form-control"
-              id="inputEmail4"
-              defaultValue={location.state.editEvents.description}
-            />
-          </div>
-          <div className=" col-md-6  ">
-            <label for="inputEmail4">Start</label>
-          </div>
-          <div className=" col-md-6 ">
-            <label for="inputEmail4">End</label>
-          </div>
-          <div className=" col-md-6 editEvent__datePicker__box">
-            <DatePicker
-              placeholderText="Start Date"
-              className="form-control"
-              selected={location.state.editEvents.start}
-              onChange={(start) =>
-                props.setNewEvent({ ...props.newEvent, start })
-              }
-            />
-          </div>
-          <div className=" col-md-6 editEvent__datePicker__box">
-            <DatePicker
-              placeholderText="End Date"
-              className="form-control"
-              selected={location.state.editEvents.end}
-              onChange={(end) => props.setNewEvent({ ...props.newEvent, end })}
-            />
-          </div>
-          <div className="col-12"></div>
-          <div className="col-12">
-            <label for="inputAddress2" className="form-label">
-              Address
-            </label>
-            <input
-              type="text"
-              className="form-control"
-              id="inputAddress2"
-              placeholder="Apartment, studio, or floor"
-            />
-          </div>
-          <Button
-            variant="primary"
-            className="editEvent__save__button"
-            onClick={redirection}
-          >
-            Save
-          </Button>{' '}
-        </form>
-      </div>
+      <EditOptions />
     </div>
   )
 }
